refactor(broker): split market_monitor.rs into directory layout (C37 of PR-3)
Pure mechanical move. market_monitor.rs (973 lines) becomes:

- market_monitor/mod.rs (~35 lines): module decls + pub use
  service::MarketMonitor + pub(crate) re-exports of the five process_*
  log-decoding helpers used by chain_monitor_v2.rs
  (process_log / process_new_logs / process_order_submitted /
  process_request_fulfilled / process_request_locked) + pub(crate) use
  types::MarketEvent.
- market_monitor/error.rs (~43 lines): MarketMonitorErr enum +
  coded_error_impl!.
- market_monitor/types.rs (~25 lines): MarketEvent enum (the decoded
  IBoundlessMarket contract events).
- market_monitor/service.rs (~942 lines): MarketMonitor<P> struct + impl,
  BrokerService impl, the five free process_* log helpers, the
  BLOCK_TIME_SAMPLE_SIZE constant, the IERC1271 sol! interface +
  ERC1271_MAGIC_VALUE constant, and the original test module.

External paths preserved:
- crate::market_monitor::MarketMonitor (broker.rs)
- crate::market_monitor::{process_log, process_new_logs,
  process_order_submitted, process_request_fulfilled,
  process_request_locked, MarketEvent} (chain_monitor_v2.rs)
