+++ crates/broker/src/market_monitor/types.rs
@@ -1,0 +1,25 @@
+// Copyright 2026 Boundless Foundation, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! Decoded market events emitted by the [`IBoundlessMarket`] contract.
+
+use boundless_market::contracts::IBoundlessMarket;
+
+/// All market event types from the [IBoundlessMarket] contract.
+#[derive(Debug, Clone)]
+pub(crate) enum MarketEvent {
+    Submitted(IBoundlessMarket::RequestSubmitted),
+    Locked(IBoundlessMarket::RequestLocked),
+    Fulfilled(IBoundlessMarket::RequestFulfilled),
+}
